No need to have two separate shapes and animations.
I really like the approach of manipulating layer properties to control
the animation timing. However, it feels a bit redundant to split up the
stroked path in two separate shapes (one to animate while you are
scrolling and another one to animate while the data is loading).
Instead you can have a single animation that is paused while scrolling
and continues once the data starts loading. This illustrates how to
manipulate the beginTime property in addition to the speed and
timeOffset properties.